+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - Kit Editor: now removes DC offset before resampling.
+ - Kit Editor: changed "Reload samples" button to "Reload sample".
 
 ## [1.9.0] - 2020-10-31
 ### Fixed